# perf: optimize entry ranges