--- a/crates/television-screen/src/results.rs
+++ b/crates/television-screen/src/results.rs
@@ -56,7 +56,7 @@ where
         // entry name
         let (entry_name, name_match_ranges) = make_matched_string_printable(
             &entry.name,
-            entry.minimal_name_match_ranges().as_deref(),
+            entry.name_match_ranges.as_deref(),
         );
         let mut last_match_end = 0;
         for (start, end) in name_match_ranges