## BEFORE (main branch)

```
results_list            time:   [12.175 µs 12.195 µs 12.217 µs]
                        change: [-1.7157% -0.9562% -0.4063%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 28 outliers among 100 measurements (28.00%)
  5 (5.00%) low severe
  14 (14.00%) low mild
  4 (4.00%) high mild
  5 (5.00%) high severe
```

## AFTER (this branch)
```
results_list            time:   [11.671 µs 11.772 µs 11.923 µs]
                        change: [-3.9850% -3.5914% -3.1779%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 28 outliers among 100 measurements (28.00%)
  22 (22.00%) low severe
  6 (6.00%) high severe
```